+++ packages/shared/ReactFeatureFlags.js
@@ -37,6 +37,9 @@ export const replayFailedUnitOfWorkWithInvokeGuardedCallback = __DEV__;
 // Warn about deprecated, async-unsafe lifecycles; relates to RFC #6:
 export const warnAboutDeprecatedLifecycles = false;
 
+// Controls enabling of experimental multi-callback support in 'schedule'
+export const scheduleModuleSupportsMultipleCallbacks = false;
+
 // Only used in www builds.
 export function addUserTimingListener() {
   invariant(false, 'Not implemented.');
