+++ packages/shared/forks/ReactFeatureFlags.native-fabric-fb.js
@@ -18,6 +18,7 @@ export const enableUserTimingAPI = __DEV__;
 export const enableGetDerivedStateFromCatch = false;
 export const warnAboutDeprecatedLifecycles = false;
 export const replayFailedUnitOfWorkWithInvokeGuardedCallback = __DEV__;
+export const scheduleModuleSupportsMultipleCallbacks = false;
 
 // React Fabric uses persistent reconciler.
 export const enableMutatingReconciler = false;
