+++ packages/shared/forks/ReactFeatureFlags.native-fb.js
@@ -26,6 +26,7 @@ export const enableUserTimingAPI = __DEV__;
 export const enableMutatingReconciler = true;
 export const enableNoopReconciler = false;
 export const enablePersistentReconciler = false;
+export const scheduleModuleSupportsMultipleCallbacks = false;
 
 // Only used in www builds.
 export function addUserTimingListener() {
