+++ packages/shared/forks/ReactFeatureFlags.native-oss.js
@@ -21,6 +21,7 @@ export const enablePersistentReconciler = false;
 export const enableUserTimingAPI = __DEV__;
 export const replayFailedUnitOfWorkWithInvokeGuardedCallback = __DEV__;
 export const warnAboutDeprecatedLifecycles = false;
+export const scheduleModuleSupportsMultipleCallbacks = false;
 
 // Only used in www builds.
 export function addUserTimingListener() {
