+++ packages/shared/forks/ReactFeatureFlags.persistent.js
@@ -18,6 +18,7 @@ export const enableUserTimingAPI = __DEV__;
 export const enableGetDerivedStateFromCatch = false;
 export const warnAboutDeprecatedLifecycles = false;
 export const replayFailedUnitOfWorkWithInvokeGuardedCallback = __DEV__;
+export const scheduleModuleSupportsMultipleCallbacks = false;
 
 // react-reconciler/persistent entry point
 // uses a persistent reconciler.
