+++ packages/shared/forks/ReactFeatureFlags.www.js
@@ -12,11 +12,12 @@ import typeof * as FeatureFlagsShimType from './ReactFeatureFlags.www';
 
 // Re-export dynamic flags from the www version.
 export const {
-  enableGetDerivedStateFromCatch,
   debugRenderPhaseSideEffects,
   debugRenderPhaseSideEffectsForStrictMode,
-  warnAboutDeprecatedLifecycles,
+  enableGetDerivedStateFromCatch,
   replayFailedUnitOfWorkWithInvokeGuardedCallback,
+  scheduleModuleSupportsMultipleCallbacks,
+  warnAboutDeprecatedLifecycles,
 } = require('ReactFeatureFlags');
 
 // The rest of the flags are static for better dead code elimination.
